services/auth_service/pyproject.toml
[project]
name = "auth_service"
version = "0.1.0"
description = "Auth microservice (owns users; issues JWTs)"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.13",
    "asyncpg>=0.30",
    "db_core",
    "fastapi[standard]>=0.115",
    "greenlet>=3.0",
    "pwdlib[argon2]>=0.2",
    "pydantic-settings>=2.5",
    "pyjwt>=2.9",
    "python-dotenv>=1.0",
    "sqlmodel>=0.0.22",
    "uvicorn>=0.30",
]

[tool.uv.sources]
db_core = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "aiosqlite>=0.20",
    "httpx>=0.27",
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
testpaths = ["tests"]
